==> doc/01-processo.Rmd <==
---
title: "Processo"
subtitle: "cas415"
author: "simona.rollet@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
```



# Processo

Nel corso del 2019, QPP laminazione ha rilevato un incremento delle caratteristiche meccaniche sulla marca interna **415M3**. Questo risulta essere problematico per alcuni clienti nel momento in cui viene trafilato in quanto si assiste ad un'ulteriore incremento di tali caratteristiche (*Rm, Rp02*) con successivo rischio di non conformità.  

A seguito di tale evento è iniziata una raccolta dei dati di processo. In questo specifico caso, si lavora su una popolazione di dati che fa riferimento alle **barre PEL**. 

I dati di processo raccolti sono relativi alle analisi chimiche, finali di colata, ed ai materiali utilizzati nella costruzione della carica solida UHP AOD. *Verranno pertanto trascurati i parametri di processo a valle del colaggio in CCO, perché rituenuti ininfluenti sulle variabili di risposta*. 

I dati di Rm ed Rp sono stati estratti da intranet e fanno riferimento alle schede testate. Il numero di schede per colata **NON** è omogeneo. 

Per riassumere i dati di processo:

+ Ogni carica UHP ha una composizione speficica in termini di rottame.

+ Ogni colata AOD eredita una carica UHP

+ Una colata AOD genera N schede
