==> doc/05-data-discovery.Rmd <==
---
title: "Data Discovery "
subtitle: "cas415"
author: "pbrunier@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
require(dplyr)
require(rlang)
require(kableExtra)
require(readr)
require(ggplot2)
source('~/dev/cas415/R/function.R')
```

leggiamo i dati 

```{r zerty}
carmet415M3 <- read_rds('/data/cas415/carmet415M3.rds')
                       
```

# Analisi della variabile risposta (Rm)

analisi Rm_mean

```{r analisi rm}
pl <- ggplot(carmet415M3)+
  geom_line(aes(index,rm))+
  geom_point(aes(index,rm, color=rm_test1))+
  geom_smooth(aes(index,rm),method = 'loess', se=FALSE, col='red')+
  geom_line(aes(index, center))+
  geom_step(aes(index, usl),color='gray',alpha=.5)+
  geom_step(aes(index, lsl),color='gray',alpha=.5)
  print(pl)
                       
```


# Analisi delle variabili input (chimica + carica)

analisi chimica nel tempo

```{r chim_temp, results='asis'}
vars <- carmet415M3 %>%
  select(sommaNbV:sfrido_molaggio_inox) %>%
  names()
titles <- paste('## Andamento nel tempo della variabile ', vars)

n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  title_i <- titles[i]
  pl <- ggplot(carmet415M3)+
    geom_line(aes(index,!!sym(var_i)))+
                geom_smooth(aes(index,!!sym(var_i)),method = 'loess', se=FALSE, col='red')
              
              cat(title_i, '\n')
              print(pl)
              cat('<p>')
}                     
```

# Analisi delle correlazioni variabili input VS variabile risposta (Rm)

analisi di correlazioni

```{r correlazioni,results='asis'}
response <- 'rm'
vars <- carmet415M3 %>% 
  select(sommaNbV:sfrido_molaggio_inox) %>%
  names()
n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  pl <- ggplot(carmet415M3) +
    geom_point(aes(!!sym(var_i),!!sym(response))) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'red',
      se = FALSE,
      method = 'loess'
    ) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'green',
      se = FALSE,
      method = 'lm'
    )
  title <- paste('## Variabile: ', var_i)
  cat(title, '\n')
  print(pl)
  cat('<p>') #necessario per terminare paragrafo
}
```




Ripeto gli stessi passaggi sul database che considera solo il valore medio di Rm per colata


leggiamo i dati 

```{r hgfd}
carmet415M3 <- read_rds('/data/cas415/carmet415M3colata.rds')
                       
```

# Analisi della variabile risposta (media di Rm per colata)

analisi Rm_mean

```{r analisi rm medio}
pl <- ggplot(carmet415M3)+
  geom_line(aes(index,rm_medio))+
  geom_point(aes(index,rm_medio, color=rm_test1))+
  geom_smooth(aes(index,rm_medio),method = 'loess', se=FALSE, col='red')+
  geom_line(aes(index, center))+
  geom_step(aes(index, usl),color='gray',alpha=.5)+
  geom_step(aes(index, lsl),color='gray',alpha=.5)
  print(pl)
                       
```



# Singola colata: Analisi delle correlazioni variabili input VS variabile risposta (media di Rm per colata)

analisi di correlazioni

```{r correlazioni Rm medio,results='asis'}
response <- 'rm_medio'
sd <- 'rm_sd'
vars <- carmet415M3 %>% 
  select(sommaNbV:sfrido_molaggio_inox) %>%
  names()
n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  pl <- ggplot(carmet415M3,aes(!!sym(var_i),!!sym(response))) +
    geom_point() +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'red',
      se = FALSE,
      method = 'loess'
    ) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'green',
      se = FALSE,
      method = 'lm'
    ) +
    geom_errorbar(aes(ymin=!!sym(response)-!!sym(sd), ymax=!!sym(response)+!!sym(sd)), width=.0001,
                 position=position_dodge(.01), colour = 'red', linetype = 'longdash') 
  title <- paste('## Variabile: ', var_i)
  cat(title, '\n')
  print(pl)
  cat('<p>') #necessario per terminare paragrafo
}
```
